- all: modified all makefiles to use "-W -Wall" as the warning   level and adapted all source files accordingly. Reactivated   the DEBUG options in all morphos makefiles aswell as the   debugging options can be dynamically disabled on the   commandline by "make DEBUG=" anyway.

+++ nlist_mcc/NList_func.h
@@ -86,7 +86,7 @@
 #define ENDQUIET                  NL_DrawQuietBG(obj,data,5,0)
 #define LESSQUIET                 NL_DrawQuietBG(obj,data,6,0)
 
-#define SetBackGround(bg)         if (bg != data->actbackground) NL_DrawQuietBG(obj,data,7,(bg));
+#define SetBackGround(bg)         if((bg) != (ULONG)data->actbackground) NL_DrawQuietBG(obj, data, 7, (bg));
 #define SetBackGroundForce(bg)    NL_DrawQuietBG(obj,data,8,(bg))
 
 #define Make_Active_Visible       NL_DrawQuietBG(obj,data,9,0)
